requirements.txt
asgiref==3.4.1
boto3==1.20.21
botocore==1.23.21
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
twilio==7.3.2
urllib3==1.26.7
whitenoise==5.3.0
